+++ packages/react-components/react-tree/src/index.ts
@@ -29,7 +29,7 @@ export {
   useTreeItemContextValues_unstable,
   useTreeItem_unstable,
 } from './TreeItem';
-export type { TreeItemProps, TreeItemState, TreeItemSlots } from './TreeItem';
+export type { TreeItemProps, TreeItemState, TreeItemSlots, TreeItemValue } from './TreeItem';
 
 export {
   TreeItemLayout,