+++ tools/generators/normalize-package-dependencies/index.ts
@@ -9,6 +9,7 @@ import {
   logger,
   createProjectGraphAsync,
   ProjectGraph,
+  readProjectConfiguration,
 } from '@nrwl/devkit';
 import * as semver from 'semver';
 
@@ -17,6 +18,10 @@ import { PackageJson } from '../../types';
 import * as chalk from 'chalk';
 
 type ProjectIssues = { [projectName: string]: { [depName: string]: string } };
+
+const NORMALIZED_INNER_WORKSPACE_VERSION = '*';
+const NORMALIZED_PRERELEASE_RANGE_VERSION = '>=9.0.0-alpha';
+const BEACHBALL_UNWANTED_PRERELEASE_RANGE_VERSION_REGEXP = /<9.0.0$/;
 
 export default async function (tree: Tree, schema: NormalizePackageDependenciesGeneratorSchema) {
   const normalizedOptions = normalizeOptions(tree, schema);
@@ -69,7 +74,7 @@ function normalizePackageJsonDependencies(tree: Tree, projectConfig: ProjectConf
 
     for (const packageName in deps) {
       if (isProjectDependencyAnWorkspaceProject(graph, packageName, projectDependencies)) {
-        const { updated } = getVersion(deps, packageName);
+        const { updated } = getVersion(tree, deps, packageName);
         deps[packageName] = updated;
       }
     }
@@ -92,18 +97,41 @@ function reportPackageJsonDependenciesIssues(issues: ProjectIssues) {
     logger.log('');
   });
 
-  logger.info(`All these dependencies version should be specified as '*'`);
+  logger.info(
+    `All these dependencies version should be specified as '${NORMALIZED_INNER_WORKSPACE_VERSION}' or '${NORMALIZED_PRERELEASE_RANGE_VERSION}' `,
+  );
   logger.info(`Fix this by running 'nx workspace-generator normalize-package-dependencies'`);
 
   throw new Error('package dependency violations found');
 }
 
-function getVersion(deps: Record<string, string>, packageName: string) {
+function getVersion(tree: Tree, deps: Record<string, string>, packageName: string) {
   const current = deps[packageName];
-  const updated = semver.prerelease(current) ? '>=9.0.0-alpha' : '*';
+  const updated = getUpdatedVersion(current);
+
   const match = current === updated;
 
   return { updated, match };
+
+  function getUpdatedVersion(currentVersion: string) {
+    if (BEACHBALL_UNWANTED_PRERELEASE_RANGE_VERSION_REGEXP.test(current)) {
+      return NORMALIZED_PRERELEASE_RANGE_VERSION;
+    }
+
+    if (currentVersion === NORMALIZED_PRERELEASE_RANGE_VERSION) {
+      const prereleasePkg = readProjectConfiguration(tree, packageName);
+      const prereleasePkgJson = readJson<PackageJson>(tree, joinPathFragments(prereleasePkg.root, 'package.json'));
+      const isPrerelease = semver.prerelease(prereleasePkgJson.version) !== null;
+
+      return isPrerelease ? NORMALIZED_PRERELEASE_RANGE_VERSION : NORMALIZED_INNER_WORKSPACE_VERSION;
+    }
+
+    if (semver.prerelease(current)) {
+      return NORMALIZED_PRERELEASE_RANGE_VERSION;
+    }
+
+    return NORMALIZED_INNER_WORKSPACE_VERSION;
+  }
 }
 
 function getPackageJsonDependenciesIssues(
@@ -133,7 +161,7 @@ function getPackageJsonDependenciesIssues(
 
     // eslint-disable-next-line guard-for-in
     for (const packageName in deps) {
-      const { match } = getVersion(deps, packageName);
+      const { match } = getVersion(tree, deps, packageName);
 
       if (isProjectDependencyAnWorkspaceProject(graph, packageName, projectDependencies) && !match) {
         issues = issues ?? {};